+++ lib/responsive_text/responsive_text.dart
@@ -21,27 +21,12 @@ class ResponsiveText extends StatelessWidget {
   final TextHeightBehavior? _textHeightBehavior;
   final Color? _selectionColor;
 
-  /// Creates a ResponsiveText widget.
-  ///
-  /// [text] is the string to display.
-  /// [style] is the [ResponsiveTextStyle] that defines how the text appears on different screen sizes.
-  /// [textAlign] defaults to [TextAlign.left].
-  /// [textDirection] specifies the directionality of the text.
-  /// [locale] is used to select a font when the same Unicode character can
-  /// be rendered differently, depending on the locale.
-  /// [softWrap] whether the text should break at soft line breaks.
-  /// [overflow] how visual overflow should be handled.
-  /// [textScaleFactor] the number of font pixels for each logical pixel.
-  /// [maxLines] an optional maximum number of lines for the text to span, wrapping if necessary.
-  /// [semanticsLabel] an alternative semantics label for this text.
-  /// [textWidthBasis] defines how to measure the width of the rendered text.
-  /// [textHeightBehavior] defines how to apply TextStyle.height to the ascent of the first line and descent of the last line.
-  /// [selectionColor] the color to use when painting the selection.
+  /// Creates a [ResponsiveText] widget.
   const ResponsiveText({
     super.key,
     required String text,
     required ResponsiveTextStyle style,
-    TextAlign textAlign = TextAlign.left, // Default value provided directly
+    TextAlign textAlign = TextAlign.left,
     TextDirection? textDirection,
     Locale? locale,
     bool? softWrap,
@@ -71,7 +56,7 @@ class ResponsiveText extends StatelessWidget {
     return Text(
       _text,
       textAlign: _textAlign,
-      style: _style.getResponsiveTextStyle(context),
+      style: _style.getResponsiveTextStyle(),
       textDirection: _textDirection,
       locale: _locale,
       softWrap: _softWrap,